feat: replace Mask Texture Editor
