+++ Editor/Inspector/RemoveMeshByMaskEditor.cs
@@ -1,12 +1,20 @@
 using Unity.Collections;
 using UnityEditor;
 using UnityEngine;
+#if AAO_MASK_TEXTURE_EDITOR
+using MaskTextureEditor = net.nekobako.MaskTextureEditor.Editor;
+#endif
 
 namespace Anatawa12.AvatarOptimizer
 {
     [CustomEditor(typeof(RemoveMeshByMask))]
     internal class RemoveMeshByMaskEditor : AvatarTagComponentEditorBase
     {
+#if AAO_MASK_TEXTURE_EDITOR
+        public const string MaskTextureEditorToken = "com.anatawa12.avatar-optimizer.remove-mesh-by-mask-editor";
+        private static readonly Vector2Int DefaultTextureSize = new(1024, 1024);
+#endif
+
         private SerializedProperty _materials = null!; // Initialized in OnEnable
         private SkinnedMeshRenderer _renderer = null!; // Initialized in OnEnable
         public bool automaticallySetWeightWhenToggle;
@@ -51,7 +59,7 @@ namespace Anatawa12.AvatarOptimizer
                         EditorGUI.indentLevel++;
                         var mask = slotConfig.FindPropertyRelative(nameof(RemoveMeshByMask.MaterialSlot.mask));
                         var mode = slotConfig.FindPropertyRelative(nameof(RemoveMeshByMask.MaterialSlot.mode));
-                        MaskTextureEditor.Inspector.DrawFields(_renderer, i, mask, mode);
+                        DrawMaskAndMode(mask, mode, _renderer, i);
                         var texture = mask.objectReferenceValue as Texture2D;
                         if (texture == null)
                         {
@@ -101,6 +109,82 @@ namespace Anatawa12.AvatarOptimizer
             }
 
             serializedObject.ApplyModifiedProperties();
+        }
+
+        private static void DrawMaskAndMode(
+            SerializedProperty mask,
+            SerializedProperty mode,
+            SkinnedMeshRenderer renderer, int slot)
+        {
+#if AAO_MASK_TEXTURE_EDITOR
+            using (new EditorGUILayout.HorizontalScope())
+            {
+                var texture = mask.objectReferenceValue as Texture2D;
+                if (texture == null)
+                {
+                    EditorGUILayout.PropertyField(mask);
+
+                    if (GUILayout.Button(AAOL10N.Tr("RemoveMeshByMask:button:create"), GUILayout.ExpandWidth(false)))
+                    {
+                        switch ((RemoveMeshByMask.RemoveMode)mode.intValue)
+                        {
+                            case RemoveMeshByMask.RemoveMode.RemoveBlack:
+                                texture = MaskTextureEditor.Utility.CreateTexture(DefaultTextureSize, Color.white);
+                                break;
+                            case RemoveMeshByMask.RemoveMode.RemoveWhite:
+                                texture = MaskTextureEditor.Utility.CreateTexture(DefaultTextureSize, Color.black);
+                                break;
+                        }
+
+                        mask.serializedObject.Update();
+                        mask.objectReferenceValue = texture;
+                        mask.serializedObject.ApplyModifiedProperties();
+
+                        if (texture != null)
+                        {
+                            MaskTextureEditor.Window.TryOpen(texture, renderer, slot, MaskTextureEditorToken);
+                        }
+
+                        // Exit GUI to avoid "EndLayoutGroup: BeginLayoutGroup must be called first."
+                        GUIUtility.ExitGUI();
+                    }
+                }
+                else
+                {
+                    var isOpenWindow = MaskTextureEditor.Window.IsOpenFor(renderer, slot, MaskTextureEditorToken);
+                    using (new EditorGUI.DisabledScope(isOpenWindow))
+                    {
+                        EditorGUILayout.PropertyField(mask);
+                    }
+
+                    var extension = System.IO.Path.GetExtension(AssetDatabase.GetAssetPath(texture));
+                    using (new EditorGUI.DisabledScope(extension != ".png"))
+                    {
+                        var shouldOpenWindow = GUILayout.Toggle(isOpenWindow,
+                            AAOL10N.Tr("RemoveMeshByMask:button:edit"),
+                            GUI.skin.button, GUILayout.ExpandWidth(false));
+                        if (isOpenWindow != shouldOpenWindow)
+                        {
+                            if (shouldOpenWindow)
+                            {
+                                MaskTextureEditor.Window.TryOpen(texture, renderer, slot, MaskTextureEditorToken);
+                            }
+                            else
+                            {
+                                MaskTextureEditor.Window.TryClose();
+                            }
+
+                            // Exit GUI to avoid "EndLayoutGroup: BeginLayoutGroup must be called first."
+                            GUIUtility.ExitGUI();
+                        }
+                    }
+                }
+            }
+#else
+            EditorGUILayout.PropertyField(mask);
+#endif
+
+            EditorGUILayout.PropertyField(mode);
         }
 
         TextureImporter? GetTextureImporter(Texture2D texture)
