--- Editor/MaskTextureEditor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 26f9d77060a8fea49a1d5f2960943360
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
