--- Editor/MaskTextureEditor/Inspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f716c26dcfc98384e862453af3342b4d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
