--- Editor/MaskTextureEditor/Shaders.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cfce59c3b6550aa4eb4d046f7dcb4445
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
