--- Editor/MaskTextureEditor/Shaders/Fill.shader
@@ -1,37 +1,0 @@
-Shader "Hidden/AvatarOptimizer/MaskTextureEditor/Fill"
-{
-    Properties
-    {
-        _Color("Color", Color) = (1.0, 1.0, 1.0, 1.0)
-    }
-
-    SubShader
-    {
-        Tags
-        {
-            "RenderType" = "Opaque"
-        }
-
-        Pass
-        {
-            CGPROGRAM
-            #pragma vertex vert
-            #pragma fragment frag
-
-            #include "UnityCG.cginc"
-
-            float4 _Color;
-
-            float4 vert(float4 v : POSITION) : SV_POSITION
-            {
-                return UnityObjectToClipPos(v);
-            }
-
-            float4 frag() : SV_Target
-            {
-                return _Color;
-            }
-            ENDCG
-        }
-    }
-}
