--- Editor/MaskTextureEditor/Shaders/Fill.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e26fa82c42ec6404b927e1a0c0f3ffe6
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
