--- Editor/MaskTextureEditor/Shaders/Inverse.shader
@@ -1,54 +1,0 @@
-Shader "Hidden/AvatarOptimizer/MaskTextureEditor/Inverse"
-{
-    Properties
-    {
-        _MainTex("Texture", 2D) = "white" {}
-    }
-
-    SubShader
-    {
-        Tags
-        {
-            "RenderType" = "Opaque"
-        }
-
-        Pass
-        {
-            CGPROGRAM
-            #pragma vertex vert
-            #pragma fragment frag
-
-            #include "UnityCG.cginc"
-
-            struct appdata
-            {
-                float4 vertex : POSITION;
-                float2 uv : TEXCOORD0;
-            };
-
-            struct v2f
-            {
-                float4 vertex : SV_POSITION;
-                float2 uv : TEXCOORD0;
-            };
-
-            sampler2D _MainTex;
-            float4 _MainTex_ST;
-
-            v2f vert(appdata v)
-            {
-                v2f o;
-                o.vertex = UnityObjectToClipPos(v.vertex);
-                o.uv = TRANSFORM_TEX(v.uv, _MainTex);
-                return o;
-            }
-
-            float4 frag(v2f i) : SV_Target
-            {
-                float4 color = tex2D(_MainTex, i.uv);
-                return float4(1.0 - color.rgb, color.a);
-            }
-            ENDCG
-        }
-    }
-}
