--- Editor/MaskTextureEditor/Shaders/Inverse.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 9690e8f230ceba845afb23f277f4ef54
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
