--- Editor/MaskTextureEditor/Shaders/Paint.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: b3225a45112500e48a551d13b88fbb76
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
