--- Editor/MaskTextureEditor/TexturePainter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62cda37c90b605d4f89417c5f85bf05e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
