--- Editor/MaskTextureEditor/TextureUndoStack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fcc4f14e9933d324a988a78d458223d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
