--- Editor/MaskTextureEditor/Window.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef74988983a72b64690935d4c44e7e54
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
