+++ Editor/com.anatawa12.avatar-optimizer.editor.asmdef
@@ -22,7 +22,8 @@
         "VRC.SDKBase",
         "UniHumanoid",
         "VRM",
-        "VRM10"
+        "VRM10",
+        "net.nekobako.MaskTextureEditor.Editor"
     ],
     "includePlatforms": [
         "Editor"
@@ -65,6 +66,11 @@
             "name": "com.vrmc.vrm",
             "expression": "",
             "define": "AAO_VRM1"
+        },
+        {
+            "name": "net.nekobako.mask-texture-editor",
+            "expression": "",
+            "define": "AAO_MASK_TEXTURE_EDITOR"
         }
     ],
     "noEngineReferences": false
